+++ packages/ui-core/src/components/Paragraph/Paragraph.test.tsx
@@ -7,10 +7,4 @@ describe('<Paragraph />', () => {
         const wrapper = shallow(<Paragraph>Text</Paragraph>);
         expect(wrapper).toMatchSnapshot();
     });
-
-    it('it renders Paragraph with div element as wrapper', () => {
-        const wrapper = shallow(<Paragraph as="div">Text</Paragraph>);
-
-        expect(wrapper).toMatchSnapshot();
-    });
 });