+++ packages/ui-core/src/components/Paragraph/Paragraph.tsx
@@ -5,33 +5,13 @@ import cnCreate from 'utils/cnCreate';
 
 interface IParagraphProps {
     /** Align text horizontally */
-    hAlign?: 'center' | 'right';
-    /** Text weight */
-    weight?: 'light' | 'regular' | 'medium' | 'bold';
-    /** Font size at all resolutions */
-    sizeAll?: 'smallest' | 'small' | 'medium' | 'large';
-    /** Font size on the resolution wide */
-    sizeWide?: 'smallest' | 'small' | 'medium' | 'large';
-    /** Font size on desktop resolution */
-    sizeDesktop?: 'smallest' | 'small' | 'medium' | 'large';
-    /** Font size on tablet resolution */
-    sizeTablet?: 'smallest' | 'small' | 'medium' | 'large';
-    /** Font size on mobile resolution */
-    sizeMobile?: 'smallest' | 'small' | 'medium' | 'large';
-    /** Vertical margin on all resolutions */
-    marginAll?: 'none' | 'small' | 'medium' | 'large' | 'largest' | 'default';
-    /** Vertical margin on the wide */
-    marginWide?: 'none' | 'small' | 'medium' | 'large' | 'largest';
-    /** Vertical margin on desktop resolution */
-    marginDesktop?: 'none' | 'small' | 'medium' | 'large' | 'largest';
-    /** Vertical padding on tablet resolution */
-    marginTablet?: 'none' | 'small' | 'medium' | 'large' | 'largest';
-    /** Vertical padding on mobile resolution */
-    marginMobile?: 'none' | 'small' | 'medium' | 'large' | 'largest';
+    align?: 'left' | 'center' | 'right';
+    /** Font options */
+    size?: 'regular' | 'small';
+    /** Vertical margin */
+    margin?: boolean;
     /** Text color */
-    color?: 'black' | 'white' | 'gray' | 'green' | 'purple' | 'red' | 'inherit';
-    /** Parent tag */
-    as?: 'p' | 'div';
+    color?: 'green' | 'purple' | 'clearWhite' | 'spbSky0' | 'spbSky1' | 'spbSky2' | 'freshAsphalt' | 'fullBlack' | 'inherit';
     /** Custom className */
     className?: string;
     children?: React.ReactNode;
@@ -40,59 +20,39 @@ interface IParagraphProps {
 const cn = cnCreate('mfui-paragraph');
 class Paragraph extends React.Component<IParagraphProps, {}> {
     static propTypes = {
-        hAlign: PropTypes.oneOf(['center', 'right']),
-        weight: PropTypes.oneOf(['light', 'regular', 'medium', 'bold']),
-        sizeAll: PropTypes.oneOf(['smallest', 'small', 'medium', 'large']),
-        sizeWide: PropTypes.oneOf(['smallest', 'small', 'medium', 'large']),
-        sizeDesktop: PropTypes.oneOf(['smallest', 'small', 'medium', 'large']),
-        sizeTablet: PropTypes.oneOf(['smallest', 'small', 'medium', 'large']),
-        sizeMobile: PropTypes.oneOf(['smallest', 'small', 'medium', 'large']),
-        marginAll: PropTypes.oneOf(['none', 'small', 'medium', 'large', 'largest', 'default']),
-        marginWide: PropTypes.oneOf(['none', 'small', 'medium', 'large', 'largest']),
-        marginDesktop: PropTypes.oneOf(['none', 'small', 'medium', 'large', 'largest']),
-        marginTablet: PropTypes.oneOf(['none', 'small', 'medium', 'large', 'largest']),
-        marginMobile: PropTypes.oneOf(['none', 'small', 'medium', 'large', 'largest']),
-        color: PropTypes.oneOf(['black', 'white', 'gray', 'green', 'purple', 'red', 'inherit']),
-        as: PropTypes.oneOf(['p', 'div']),
+        align: PropTypes.oneOf(['left', 'center', 'right']),
+        size: PropTypes.oneOf(['regular', 'small']),
+        margin: PropTypes.bool,
+        color: PropTypes.oneOf(['green', 'purple', 'clearWhite', 'spbSky0', 'spbSky1', 'spbSky2', 'freshAsphalt', 'fullBlack', 'inherit']),
         children: PropTypes.node,
     };
 
     static defaultProps: Partial<IParagraphProps> = {
-        marginAll: 'medium',
-        sizeAll: 'medium',
-        color: 'black',
-        weight: 'regular',
-        as: 'p',
+        margin: true,
+        size: 'regular',
+        color: 'freshAsphalt',
     };
 
     render() {
-        const TagName = this.props.as as React.ElementType;
         const {
-            sizeAll, sizeWide, sizeDesktop, sizeTablet, sizeMobile,
-            marginAll, marginWide, marginDesktop, marginTablet,
-            marginMobile, hAlign, color, weight, className, children,
+            size,
+            align,
+            color,
+            margin,
+            children,
+            className,
         } = this.props;
-        const isMarginAll = marginAll === 'default' ? false : marginAll;
 
         return (
-            <TagName
+            <p
                 className={cn('', {
-                    'size-all': sizeAll,
-                    'size-wide': sizeWide,
-                    'size-desktop': sizeDesktop,
-                    'size-tablet': sizeTablet,
-                    'size-mobile': sizeMobile,
-                    'margin-all': isMarginAll,
-                    'margin-wide': marginWide,
-                    'margin-desktop': marginDesktop,
-                    'margin-tablet': marginTablet,
-                    'margin-mobile': marginMobile,
-                    'h-align': hAlign,
-                    color: color,
-                    weight: weight,
+                    size,
+                    align,
+                    color,
+                    margin,
                 }, className)}>
                 {children}
-            </TagName>
+            </p>
         );
     }
 }