+++ packages/ui-core/CHANGELOG.md
@@ -1,8 +1,6 @@
 # Changelog
 
 ## Unreleased
-- **Accordion**
-    - fixed title style, text padding, border-bottom in opened accordion
 - **Preloader**
     - added new component
     - fixed tslint errors
@@ -23,6 +21,7 @@
     - changed the parameter name from `lg` in the `innerPadding` property to `default`
 - **Accordion**
     - add Accordion component
+    - fixed title style, text padding, border-bottom in opened accordion
 - **Paragraph**
     - update with a new guide
     - update test